walkViews refactored into a mixin and added to destroy-submisison component
