+++ app/templates/components/destroy-submission.hbs
@@ -2,7 +2,7 @@
   {{loading-spinner}}
 {{else}}
   <button
-    {{action 'destroy' target='view.parentView'}}
+    {{action 'destroy'}}
     class="button-secondary"
     data-test="button-destroy"
   >
