+++ tests/dummy/app/resources/posts/new/template.hbs
@@ -16,3 +16,7 @@
   <dt>Description</dt>
   <dd>{{description}}</dd>
 </dl>
+
+{{#if editing}}
+  {{destroy-submission}}
+{{/if}}
