+++ tests/dummy/app/templates/application.hbs
@@ -7,6 +7,7 @@
 <ul>
   <li>{{#link-to 'posts'}}Posts{{/link-to}}</li>
   <li>{{#link-to 'posts.new'}}New post{{/link-to}}</li>
+  <li>{{#link-to 'post.edit' 1}}Edit post{{/link-to}}</li>
 </ul>
 
 {{outlet}}
